+++ src/components/Header/Header.js
@@ -3,8 +3,10 @@ import styled from 'styled-components'
 
 import Section from '../layout/Section'
 import AnimatedShapes from '../AnimatedShapes/AnimatedShapes'
-import SearchForm from './SearchForm'
-import MenuBar from './MenuBar'
+import SearchForm from './SearchForm/SearchForm'
+import MenuBar from './MenuBar/MenuBar'
+import HeaderContent from './styles/HeaderContent'
+import FormContainer from './styles/FormContainer'
 
 const Header = (props) => {
   return(
@@ -33,35 +35,8 @@ const Header = (props) => {
   )
 }
 
-const HeaderContent = styled.div`
-  height: ${({theme}) => theme.layout.headerHight}px;
-  padding-top: 65px;
-  justify-content: flex-start;
-  align-items: center;
-  display: flex;
-  overflow-x: visible;
-  flex-direction: column;
-  max-width: 800px;
-  margin: 0 auto;
-  position: relative;
-
-  h1 {
-    margin-bottom: 30px;
-    font-size: 4rem;
-  }
-  
-  p {
-    max-width: 400px;
-  }
-`
-
 const BannerShapes = styled(AnimatedShapes)`
   top: -30px;
-`
-
-const FormContainer = styled.div`
-  position: absolute;
-  bottom: 80px;
 `
 
 export default Header
