+++ src/components/Header/styles/FormContainer.js
@@ -1,0 +1,12 @@
+import styled from 'styled-components'
+
+const FormContainer = styled.div`
+  position: absolute;
+  bottom: 80px;
+  overflow: hidden;
+  width: 100%;
+  display; flex;
+  justify-content: center;
+`
+
+export default FormContainer
