+++ src/index.js
@@ -3,8 +3,8 @@ import ReactDOM from 'react-dom'
 import { ThemeProvider } from 'styled-components'
 import { BrowserRouter } from 'react-router-dom'
 
-import theme from './styles/theme'
-import GlobalStyle from './styles/global'
+import theme from './globalstyles/theme'
+import GlobalStyle from './globalstyles/global'
 
 import UserContextProvider from './context/user/UserContext'
 import ResultsContextProvider from './context/results/ResultsContext'
